Refactor Python bindings
This commit refactors and reorganizes the python
bindings to improve readability and maintanace.
