+++ imgal_python/src/child_modules/mod.rs
@@ -1,0 +1,5 @@
+pub mod integration_module;
+pub mod parameters_module;
+pub mod phasor_module;
+pub mod simulation_module;
+pub mod statistics_module;
