+++ imgal_python/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod child_modules;
 pub mod functions;
 pub mod parent_module;
+mod utils;
